requirements.txt
Pillow==10.0.1
requests==2.31.0 